Added different left and right servo positions
Now you can adjust the servo positions individually of left and right instead of having both servos at the same value (just in case both servos are not at 90 degrees at the same value). Also commented out the sweeping of servo positions for the left and right turns and just went straight to the position needed. (This is because the delay for left and right turns may keep the turn going longer than necessary)